+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 38
+#define BUILD 39